+++ src/SMAPI.Web/wwwroot/Content/js/index.js
@@ -1,8 +1,11 @@
 $(document).ready(function () {
     var pufferchick = $("#pufferchick");
-    $(".download").hover(function () {
-        pufferchick.attr("src", "Content/images/pufferchick-cool.png");
-    }, function () {
-        pufferchick.attr("src", "favicon.ico");
-    });
+    $(".download").hover(
+        function () {
+            pufferchick.attr("src", "Content/images/pufferchick-cool.png");
+        },
+        function () {
+            pufferchick.attr("src", "favicon.ico");
+        }
+    );
 });